backend/requirements.txt
fastapi
uvicorn
pandas
python-dotenv
sqlalchemy
psycopg2-binary
strawberry-graphql==0.123.0

scipy
tiktoken
xai-sdk